requirements.txt
datasets
transformers
pandas
numpy
scikit-learn
faiss-cpu
openai
python-dotenv
rank-bm25
tqdm
streamlit
matplotlib
seaborn
google-generativeai
anthropic
